Hw2

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -1,46 +1,82 @@
 package ru.javawebinar.topjava.repository.inmemory;
 
+import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
+import ru.javawebinar.topjava.util.DateTimeUtil;
 import ru.javawebinar.topjava.util.MealsUtil;
 
+import java.time.LocalDateTime;
 import java.util.Collection;
+import java.util.Comparator;
 import java.util.Map;
+import java.util.Objects;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Predicate;
+import java.util.stream.Collectors;
 
+@Repository
 public class InMemoryMealRepository implements MealRepository {
     private final Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private final AtomicInteger counter = new AtomicInteger(0);
 
     {
-        MealsUtil.meals.forEach(this::save);
+        MealsUtil.meals.forEach(m -> save(m, 1));
     }
 
     @Override
-    public Meal save(Meal meal) {
+    public Meal save(Meal meal, Integer userId) {
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
             repository.put(meal.getId(), meal);
             return meal;
         }
+        Meal existing = repository.get(meal.getId());
+        if (existing == null || !mealBelongsToUser(userId, existing)) return null;
         // handle case: update, but not present in storage
         return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
     }
 
     @Override
-    public boolean delete(int id) {
+    public boolean delete(int id, Integer userId) {
+        Meal meal = repository.get(id);
+        if (meal == null || !mealBelongsToUser(userId, meal)) {
+            return false;
+        }
         return repository.remove(id) != null;
     }
 
     @Override
-    public Meal get(int id) {
-        return repository.get(id);
+    public Meal get(int id, Integer userId) {
+        Meal meal = repository.get(id);
+        if (meal == null || !mealBelongsToUser(userId, meal)) {
+            return null;
+        }
+        return meal;
     }
 
     @Override
-    public Collection<Meal> getAll() {
-        return repository.values();
+    public Collection<Meal> getAll(Integer userId) {
+        return getAll(m -> mealBelongsToUser(userId, m));
+    }
+
+    @Override
+    public Collection<Meal> getAll(Integer userId, LocalDateTime startDateTime, LocalDateTime endDateTime) {
+        return getAll(m -> mealBelongsToUser(userId, m)
+                && DateTimeUtil.isBetweenHalfOpen(m.getDateTime(), startDateTime, endDateTime));
+    }
+
+    private Collection<Meal> getAll(Predicate<Meal> predicate) {
+        return repository.values()
+                .stream()
+                .filter(predicate)
+                .sorted(((Comparator<Meal>) (o1, o2) -> o1.getDateTime().compareTo(o2.getDateTime())).reversed())
+                .collect(Collectors.toList());
+    }
+
+    private boolean mealBelongsToUser(Integer userId, Meal existing) {
+        return Objects.equals(existing.getUserId(), userId);
     }
 }
